# docs/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: THE C1RCLE Frontend Integration API
  version: 0.1.0-draft
  description: |
    SUPERSEDED for wire shapes (2026-08-29). This draft uses `/session` and a
    `{ data, meta }` / `{ success, error }` envelope taken from the frozen V2
    manifest. The gateway ships something different: `/api/v2/auth/session`,
    a BARE-DTO success body, and a FLAT error envelope
    `{ code, message, status, requestId, fieldErrors? }` (lowercase `code`).
    The authoritative machine-readable contract is `packages/contracts`
    (zod schemas, shared by backend and frontend); see
    `docs/api-contracts/{auth-and-permissions,error-contract}.md` for the
    corrected prose. Regenerate this file from `packages/contracts` before
    relying on it.

    Route activation is governed by `apps/api-gateway/src/routes/v2/route-manifest.ts`
    (the runtime code, not a doc). Do not expose Firestore documents or
    provider payloads.
servers:
  - url: https://api.example.invalid/api/v2
    description: Replace with the approved staging/production API base URL.
security:
  - bearerAuth: []
...
paths:
...
  /orders/{orderId}/payment-intents:
    post:
      tags: [Orders]
      summary: Create a provider payment action for an order
      operationId: createPaymentIntent
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '201':
          description: Payment action
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentIntentResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /payments/{paymentId}/confirm:
    post:
      tags: [Orders]
      summary: Confirm a provider payment after backend verification
      operationId: confirmPayment
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: paymentId, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerPaymentId, providerOrderId, providerSignature]
              properties:
                providerPaymentId: { type: string }
                providerOrderId: { type: string }
                providerSignature: { type: string }
      responses:
        '200':
          description: Verified payment and current order state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: FirebaseIDToken
      description: Approved session cookie may be used for web clients instead.
  parameters:
...
    IdempotencyKey:
      { name: Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } }
...
    OrderId: { name: orderId, in: path, required: true, schema: { type: string } }
  responses:
    Unauthenticated:
      description: Missing, expired, or revoked session
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Forbidden:
      description: Authenticated but not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: Resource does not exist or is not visible
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Conflict:
      description: State, inventory, version, or idempotency conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Validation:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  schemas:
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string, format: uuid }
        nextCursor: { type: string, nullable: true }
    ErrorDetail:
      type: object
      required: [message]
      properties:
        path: { type: string }
        message: { type: string }
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: array, items: { $ref: '#/components/schemas/ErrorDetail' } }
        requestId: { type: string, format: uuid }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: '#/components/schemas/ApiError' }
...
    Money:
      type: object
      required: [amountMinor, currency]
      properties:
        amountMinor: { type: integer, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
...
    Order:
      type: object
      required: [id, status, total, createdAt]
      properties:
        id: { type: string }
        status:
          { type: string, enum: [PENDING_PAYMENT, PAID, FAILED, CANCELLED, EXPIRED, REFUNDED] }
        total: { $ref: '#/components/schemas/Money' }
        createdAt: { type: string, format: date-time }
    PaymentIntent:
      type: object
      required: [id, provider, providerOrderId, amount, status]
      properties:
        id: { type: string }
        provider: { type: string, enum: [razorpay] }
        providerOrderId: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [CREATED, PROCESSING, SUCCEEDED, FAILED] }
        clientOptions: { type: object, additionalProperties: true }
...
    OrderResponse:
      type: object
      required: [data, meta]
      properties:
        {
          data: { $ref: '#/components/schemas/Order' },
          meta: { $ref: '#/components/schemas/Meta' },
        }
    PaymentIntentResponse:
      type: object
      required: [data, meta]
      properties:
        {
          data: { $ref: '#/components/schemas/PaymentIntent' },
          meta: { $ref: '#/components/schemas/Meta' },
        }
...